Restaurado el dev_seeds.rb original, cambiado el nombre del archivo de seeds personalizado a custom_seeds.rb y añadida una tarea rake para poder cargar archivos de seeds personalizados.
